LPS-125698 commerce-account-service: (unrelated) moves upgrade process to proper package
