+++ modules/apps/commerce/commerce-account-service/src/main/java/com/liferay/commerce/account/internal/upgrade/CommerceAccountUpgradeStepRegistrator.java
@@ -21,7 +21,7 @@ import com.liferay.commerce.account.internal.upgrade.v1_2_0.CommerceAccountGroup
 import com.liferay.commerce.account.internal.upgrade.v1_3_0.CommerceAccountNameUpgradeProcess;
 import com.liferay.commerce.account.internal.upgrade.v1_4_0.CommerceAccountDefaultAddressesUpgradeProcess;
 import com.liferay.commerce.account.internal.upgrade.v2_0_0.CommerceAccountGroupSystemUpgradeProcess;
-import com.liferay.commerce.account.internal.upgrade.v3_0_0.CommerceAccountOrganizationRelUpgradeProcess;
+import com.liferay.commerce.account.internal.upgrade.v4_0_0.CommerceAccountOrganizationRelUpgradeProcess;
 import com.liferay.commerce.account.internal.upgrade.v5_0_0.CommerceAccountUserRelUpgradeProcess;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
